refactor(types): single-home the search API contract in the SDK
The 14-type search cluster (SearchSchemaField, SearchIndexDefinition,
SearchContentTypeDefinition, SearchQueryParams, SearchResult,
SearchResponse, SearchHealth{Index,Table,Snapshot}, SearchTransformOp,
FilePatternMapper, FileBackedContentTypeDefinition, SearchAPI,
SearchMaintenanceAPI) was duplicated across core and the SDK. Two real
drifts where core was the accurate superset, now reconciled UP into the
canonical SDK declaration: the SDK lacked SearchMaintenanceAPI +
SearchAPI.maintenance (used by the memory plugin via ctx.search) and its
SearchHealthSnapshot inlined a narrower per-table shape without
indexHealth/SearchHealthIndex. Core now re-exports the cluster. Audit
finding: sdk-gaps.
